backend_django/requirements.txt
Django>=5.0.0
django-environ
django-cors-headers
django-filter
django-taggit
djangorestframework
dj-database-url

psycopg2-binary
google-generativeai
python-dotenv
whitenoise
gunicorn
celery
redis
numpy
requests
pytest
pytest-django
django-unfold
django-simple-history
djangorestframework-simplejwt
